refactor: deleted unused files and libs
