--- app/src/main/java/com/kleckus/chucknorrisfacts/system/SysEnvironment.kt
@@ -1,5 +1,0 @@
-package com.kleckus.chucknorrisfacts.system
-
-enum class Environment{
-    PRODUCTION, TESTING
-}
